Dianka

+++ migrate-cts.mjs
@@ -18,18 +18,22 @@ const handleSlicezone = (sz) => {
       ...sz.config,
       choices: Object.fromEntries(
         Object.entries(sz.config.choices).map(([key, val]) => {
-          fs.mkdirSync(`./slices/${key}`)
-          fs.writeFileSync(
-            `./slices/${key}/model.json`,
-            JSON.stringify(
-              {
-                ...migrate.default(val, { sliceName: key, from: 'slices'}, { cwd: './' }).model,
-                migrated: undefined
-              },
-              null,
-              2
+          if (!fs.existsSync(`./slices/${key}`)) {
+            fs.mkdirSync(`./slices/${key}`)
+            fs.writeFileSync(
+              `./slices/${key}/model.json`,
+              JSON.stringify(
+                {
+                  ...migrate.default(val, { sliceName: key, from: 'slices' }, { cwd: './' }).model,
+                  migrated: undefined
+                },
+                null,
+                2
+              )
             )
-          )
+          } else {
+            console.log("slice", key, "already exists")
+          }
           return [key, { type: "SharedSlice" }]
         })
       )
@@ -37,59 +41,63 @@ const handleSlicezone = (sz) => {
   }
 }
 
-;(() => {
-fetch('https://customtypes.prismic.io/customtypes', {
-  headers: HEADERS
-}).then(res => res.json())
-.then(cts => {
-  fs.mkdirSync('./customtypes')
-  fs.mkdirSync('./slices')
-  cts.forEach((ct) => {
+  ; (() => {
+    fetch('https://customtypes.prismic.io/customtypes', {
+      headers: HEADERS
+    })
+      .then(res => res.json())
+      .then(cts => {
+        let str = JSON.stringify(cts);
+        str = str.replaceAll("\"body\":", "\"slices\":");
+        cts = JSON.parse(str);
+        fs.mkdirSync('./customtypes')
+        fs.mkdirSync('./slices')
+        cts.forEach((ct) => {
 
-    const newCt = {
-      ...ct,
-      json: (() => {
-        return Object.fromEntries(
-          Object.entries(ct.json).map(([tabKey, tab]) => {
+          const newCt = {
+            ...ct,
+            json: (() => {
+              return Object.fromEntries(
+                Object.entries(ct.json).map(([tabKey, tab]) => {
 
-            return [
-              tabKey,
-              Object.fromEntries(
-                Object.entries(tab).map(([fieldKey, field]) => {
-                  if (field.type === 'Slices') {
-                    return [fieldKey, handleSlicezone(field)]
-                  }
-                  return [fieldKey, field]
+                  return [
+                    tabKey,
+                    Object.fromEntries(
+                      Object.entries(tab).map(([fieldKey, field]) => {
+                        if (field.type === 'Slices') {
+                          return [fieldKey, handleSlicezone(field)]
+                        }
+                        return [fieldKey, field]
+                      })
+                    )
+                  ]
                 })
               )
-            ]
-          })
-        )
-      })()
-    }
+            })()
+          }
 
-    fs.mkdirSync(`./customtypes/${newCt.id}`)
-    fs.writeFileSync(`./customtypes/${newCt.id}/index.json`, JSON.stringify(newCt, null, 2))
+          fs.mkdirSync(`./customtypes/${newCt.id}`)
+          fs.writeFileSync(`./customtypes/${newCt.id}/index.json`, JSON.stringify(newCt, null, 2))
 
-    // const tabs = Object.entries(ct.json)
-    // const newTabs = tabs.map(t => {
-    //   const [tabKey, tab] = t
-    //   const entries = Object.entries(tab)
-    //   console.log({ entries, tab })
+          // const tabs = Object.entries(ct.json)
+          // const newTabs = tabs.map(t => {
+          //   const [tabKey, tab] = t
+          //   const entries = Object.entries(tab)
+          //   console.log({ entries, tab })
 
-    //   const newTabEntries = entries.map((entry) => {
-    //     const [key, val] = entry
-    //     if (val.type === 'Slices') {
-    //       console.log({ val })
-    //     }
-    //     return [key, val]
-    //   })
-    //   return newTabEntries
-    // })
-    // console.log({
-    //   ...ct,
-    //   json: JSON.stringify(Object.fromEntries(newTabs))
-    // })
-  })
-})
-})()
+          //   const newTabEntries = entries.map((entry) => {
+          //     const [key, val] = entry
+          //     if (val.type === 'Slices') {
+          //       console.log({ val })
+          //     }
+          //     return [key, val]
+          //   })
+          //   return newTabEntries
+          // })
+          // console.log({
+          //   ...ct,
+          //   json: JSON.stringify(Object.fromEntries(newTabs))
+          // })
+        })
+      })
+  })()
